✨ feat(PartDialog): implement part creation and editing dialog
- Introduce PartDialog component for creating and editing parts with detailed level management
- Enhance form handling to support dynamic level assignments and building selections
- Improve user experience with responsive design and clear labeling of input fields
- Integrate dialog with PartsTab for seamless part management functionality
