+++ src/app/dashboard/(registre-securite)/sites/[siteId]/components/PartsTab.tsx
@@ -19,30 +19,19 @@ import {
   CardActions,
   CardContent,
   Chip,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  FormControl,
-  FormControlLabel,
   Grid,
   IconButton,
-  InputAdornment,
-  InputLabel,
-  MenuItem,
   Paper,
-  Select,
-  Switch,
   Table,
   TableBody,
   TableCell,
   TableContainer,
   TableHead,
   TableRow,
-  TextField,
   Typography
 } from '@mui/material';
 import React, { useEffect, useState } from 'react';
+import PartDialog from './PartDialog';
 
 interface PartsTabProps {
   siteId: number;
@@ -54,10 +43,6 @@ interface Filters {
   search: string;
   buildingId: number | '';
   includeDeleted: boolean;
-}
-
-interface PartFormData extends CreatePartDto {
-  levelCount: number;
 }
 
 const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = false }) => {
@@ -79,19 +64,6 @@ const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = 
   const [dialogOpen, setDialogOpen] = useState(false);
   const [dialogMode, setDialogMode] = useState<'create' | 'edit'>('create');
   const [selectedPart, setSelectedPart] = useState<Part | null>(null);
-
-  // Form states
-  const [formData, setFormData] = useState<PartFormData>({
-    name: '',
-    buildingId: 0,
-    levelCount: 1,
-    type: 'PRIVATE',
-    isIcpe: false,
-    erpTypeCodes: ['J'],
-  });
-
-  // Level assignment states
-  const [levelAssignments, setLevelAssignments] = useState<LevelAssignment[]>([]);
 
   // Load data
   useEffect(() => {
@@ -173,129 +145,28 @@ const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = 
     setFilters(prev => ({ ...prev, [field]: value }));
   };
 
-  // Initialize level assignments based on levelCount
-  const initializeLevelAssignments = (levelCount: number, existingPart?: Part) => {
-    const assignments: LevelAssignment[] = [];
-    for (let i = 1; i <= levelCount; i++) {
-      assignments.push({
-        levelNumber: i,
-        buildingFloorId: null,
-        partFloorData: {
-          name: `Niveau ${i}`,
-          publicCount: 0,
-          staffCount: 0,
-          exploitationSurface: 0,
-          glaSurface: 0,
-          publicAccessSurface: 0,
-          levelNumber: i,
-        }
-      });
-    }
-
-    // If editing, try to populate with existing data
-    if (existingPart?.partFloors) {
-      existingPart.partFloors.forEach((partFloor, index) => {
-        if (assignments[index]) {
-          assignments[index].buildingFloorId = partFloor.buildingFloor?.id || null;
-          assignments[index].partFloorData = {
-            name: partFloor.name,
-            publicCount: partFloor.publicCount,
-            staffCount: partFloor.staffCount,
-            exploitationSurface: partFloor.exploitationSurface,
-            glaSurface: partFloor.glaSurface,
-            publicAccessSurface: partFloor.publicAccessSurface,
-            levelNumber: partFloor.levelNumber,
-          };
-        }
-      });
-    }
-
-    setLevelAssignments(assignments);
-  };
-
   const openCreateDialog = () => {
     setDialogMode('create');
     setSelectedPart(null);
-    const availableBuildings = buildings.filter(b => !b.deletedAt);
-    setFormData({
-      name: '',
-      buildingId: availableBuildings.length > 0 ? availableBuildings[0].id : 0,
-      levelCount: 1,
-      type: 'PRIVATE',
-      isIcpe: false,
-      erpTypeCodes: ['J'],
-    });
-    initializeLevelAssignments(1);
     setDialogOpen(true);
   };
 
   const openEditDialog = (part: Part) => {
     setDialogMode('edit');
     setSelectedPart(part);
-    setFormData({
-      name: part.name,
-      buildingId: part.building?.id || 0,
-      levelCount: part.partFloors?.length || 1,
-      type: part.type || 'PRIVATE',
-      isIcpe: part.isIcpe,
-      erpTypeCodes: part.erpTypes || ['J'],
-    });
-    initializeLevelAssignments(part.partFloors?.length || 1, part);
     setDialogOpen(true);
   };
 
   const closeDialog = () => {
     setDialogOpen(false);
     setSelectedPart(null);
-    setLevelAssignments([]);
   };
 
-  const handleFormChange = (field: keyof PartFormData, value: any) => {
-    setFormData(prev => ({ ...prev, [field]: value }));
-
-    // If levelCount changes, reinitialize level assignments
-    if (field === 'levelCount') {
-      initializeLevelAssignments(value, selectedPart || undefined);
-    }
-  };
-
-  const updateLevelAssignment = (levelNumber: number, field: keyof LevelAssignment, value: any) => {
-    setLevelAssignments(prev =>
-      prev.map(assignment =>
-        assignment.levelNumber === levelNumber
-          ? { ...assignment, [field]: value }
-          : assignment
-      )
-    );
-  };
-
-  const updatePartFloorData = (levelNumber: number, field: string, value: any) => {
-    setLevelAssignments(prev =>
-      prev.map(assignment =>
-        assignment.levelNumber === levelNumber
-          ? {
-            ...assignment,
-            partFloorData: {
-              ...assignment.partFloorData!,
-              [field]: value
-            }
-          }
-          : assignment
-      )
-    );
-  };
-
-  const getAvailableBuildingFloors = (selectedBuildingId: number) => {
-    return buildingFloors
-      .filter(floor => floor.building.id === selectedBuildingId)
-      .filter(floor => filters.includeDeleted || !floor.deletedAt);
-  };
-
-  const handleSubmit = async () => {
+  const handleDialogSubmit = async (formData: CreatePartDto | UpdatePartDto, levelAssignments: LevelAssignment[], partId?: number) => {
     try {
       if (dialogMode === 'create') {
         // Create the part first and get the created part with its ID
-        const createdPart = await partService.createPart(formData);
+        const createdPart = await partService.createPart(formData as CreatePartDto);
 
         // Create part floors for each level assignment
         for (const assignment of levelAssignments) {
@@ -309,15 +180,9 @@ const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = 
         }
 
         onNotification('Partie créée avec succès', 'success');
-      } else if (selectedPart) {
+      } else if (partId) {
         // Update the part
-        const updateData: UpdatePartDto = {
-          name: formData.name,
-          type: formData.type,
-          isIcpe: formData.isIcpe,
-          erpTypeCodes: formData.erpTypeCodes,
-        };
-        await partService.updatePart(selectedPart.id, updateData);
+        await partService.updatePart(partId, formData as UpdatePartDto);
 
         // TODO: Handle updating part floors - this requires more complex logic
         // to determine which ones to create, update, or delete
@@ -325,11 +190,11 @@ const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = 
         onNotification('Partie modifiée avec succès', 'success');
       }
 
-      closeDialog();
       loadData();
     } catch (error) {
       console.error('Error saving part:', error);
       onNotification('Erreur lors de la sauvegarde', 'error');
+      throw error; // Re-throw to let dialog handle loading state
     }
   };
 
@@ -585,426 +450,17 @@ const PartsTab: React.FC<PartsTabProps> = ({ siteId, onNotification, disabled = 
         </>
       )}
 
-      {/* Create/Edit Dialog */}
-      <Dialog
+      {/* Part Dialog */}
+      <PartDialog
         open={dialogOpen}
-        onClose={(event, reason) => {
-          if (reason !== 'backdropClick' && reason !== 'escapeKeyDown') {
-            closeDialog();
-          }
-        }}
-        maxWidth="lg"
-        fullWidth
-        disableEscapeKeyDown
-      >
-        <DialogTitle>
-          {dialogMode === 'create' ? 'Créer une partie' : 'Modifier la partie'}
-        </DialogTitle>
-        <DialogContent>
-          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 3, pt: 1 }}>
-            {/* Basic Information */}
-            <Box sx={{ display: 'grid', gridTemplateColumns: { xs: '1fr', sm: '1fr 1fr' }, gap: 2 }}>
-              <TextField
-                label="Nom de la partie"
-                value={formData.name}
-                onChange={(e) => handleFormChange('name', e.target.value)}
-                required
-                fullWidth
-              />
-
-              <FormControl required fullWidth>
-                <InputLabel>Bâtiment</InputLabel>
-                <Select
-                  value={formData.buildingId}
-                  onChange={(e) => handleFormChange('buildingId', Number(e.target.value))}
-                  label="Bâtiment"
-                >
-                  {buildings
-                    .filter(building => filters.includeDeleted || !building.deletedAt)
-                    .map((building) => (
-                      <MenuItem key={building.id} value={building.id}>
-                        {building.name}
-                      </MenuItem>
-                    ))}
-                </Select>
-              </FormControl>
-            </Box>
-
-            <Box sx={{ display: 'grid', gridTemplateColumns: { xs: '1fr', sm: '1fr 1fr 1fr' }, gap: 2 }}>
-              <TextField
-                label="Nombre de niveaux"
-                type="number"
-                value={formData.levelCount}
-                onChange={(e) => handleFormChange('levelCount', Math.max(1, parseInt(e.target.value) || 1))}
-                inputProps={{ min: 1, max: 20 }}
-                required
-                fullWidth
-              />
-
-              <FormControl required fullWidth>
-                <InputLabel>Type</InputLabel>
-                <Select
-                  value={formData.type}
-                  onChange={(e) => handleFormChange('type', e.target.value)}
-                  label="Type"
-                >
-                  <MenuItem value="PRIVATE">Privée</MenuItem>
-                  <MenuItem value="COMMUNAL">Commune</MenuItem>
-                </Select>
-              </FormControl>
-
-              <FormControl fullWidth>
-                <InputLabel>Code ERP</InputLabel>
-                <Select
-                  value={formData.erpTypeCodes}
-                  onChange={(e) => handleFormChange('erpTypeCodes', e.target.value)}
-                  label="Code ERP"
-                >
-                  <MenuItem value="J">J - Structures d&apos;accueil pour personnes âgées</MenuItem>
-                  <MenuItem value="L">L - Salles de spectacles</MenuItem>
-                  <MenuItem value="M">M - Magasins de vente</MenuItem>
-                  <MenuItem value="N">N - Restaurants et débits de boissons</MenuItem>
-                  <MenuItem value="O">O - Hôtels et pensions de famille</MenuItem>
-                  <MenuItem value="P">P - Salles de danse et salles de jeux</MenuItem>
-                  <MenuItem value="R">R - Établissements de soins</MenuItem>
-                  <MenuItem value="S">S - Bibliothèques, centres de documentation</MenuItem>
-                  <MenuItem value="T">T - Salles d&apos;exposition</MenuItem>
-                  <MenuItem value="U">U - Établissements de soins avec hébergement</MenuItem>
-                  <MenuItem value="V">V - Établissements de culte</MenuItem>
-                  <MenuItem value="W">W - Administrations, banques, bureaux</MenuItem>
-                  <MenuItem value="X">X - Établissements sportifs couverts</MenuItem>
-                  <MenuItem value="Y">Y - Musées</MenuItem>
-                </Select>
-              </FormControl>
-            </Box>
-
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={formData.isIcpe}
-                  onChange={(e) => handleFormChange('isIcpe', e.target.checked)}
-                />
-              }
-              label="Installation Classée pour la Protection de l'Environnement (ICPE)"
-            />
-
-            {/* Level Assignments */}
-            {levelAssignments.length > 0 && (
-              <Box>
-                <Typography variant="h6" sx={{ mb: 2 }}>
-                  Affectation des niveaux aux étages de bâtiment
-                </Typography>
-
-                {/* Desktop View */}
-                <Box sx={{ display: { xs: 'none', md: 'block' } }}>
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                    {levelAssignments.map((assignment) => (
-                      <Paper key={assignment.levelNumber} variant="outlined" sx={{ p: 2 }}>
-                        <Typography variant="subtitle1" fontWeight="medium" sx={{ mb: 2 }}>
-                          Niveau {assignment.levelNumber}
-                        </Typography>
-
-                        <Box sx={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fit, minmax(200px, 1fr))', gap: 2 }}>
-                          <FormControl size="small" fullWidth>
-                            <InputLabel>Étage de bâtiment</InputLabel>
-                            <Select
-                              label="Étage de bâtiment"
-                              value={assignment.buildingFloorId || ''}
-                              onChange={(e) => updateLevelAssignment(
-                                assignment.levelNumber,
-                                'buildingFloorId',
-                                e.target.value === '' ? null : Number(e.target.value)
-                              )}
-                              displayEmpty
-                            >
-                              <MenuItem value="">Non affecté</MenuItem>
-                              {getAvailableBuildingFloors(formData.buildingId).map((floor) => (
-                                <MenuItem key={floor.id} value={floor.id}>
-                                  {floor.name}
-                                </MenuItem>
-                              ))}
-                            </Select>
-                          </FormControl>
-
-                          <TextField
-                            label="Nom de l'étage de partie"
-                            size="small"
-                            value={assignment.partFloorData?.name || ''}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'name',
-                              e.target.value
-                            )}
-                            placeholder={`Niveau ${assignment.levelNumber}`}
-                            fullWidth
-                          />
-
-                          <TextField
-                            label="Niveau de partie"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.levelNumber || assignment.levelNumber}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'levelNumber',
-                              parseInt(e.target.value) || assignment.levelNumber
-                            )}
-                            inputProps={{ min: 0 }}
-                            fullWidth
-                          />
-
-                          <TextField
-                            label="Public"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.publicCount || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'publicCount',
-                              parseInt(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0 }}
-                            fullWidth
-                          />
-
-                          <TextField
-                            label="Personnel"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.staffCount || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'staffCount',
-                              parseInt(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0 }}
-                            fullWidth
-                          />
-
-                          <TextField
-                            label="Surface exploitation"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.exploitationSurface || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'exploitationSurface',
-                              parseFloat(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0, step: 0.1 }}
-                            fullWidth
-                            slotProps={{
-                              input: {
-                                endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                              }
-                            }}
-                          />
-
-                          <TextField
-                            label="Surface GLA"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.glaSurface || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'glaSurface',
-                              parseFloat(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0, step: 0.1 }}
-                            fullWidth
-                            slotProps={{
-                              input: {
-                                endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                              }
-                            }}
-                          />
-
-                          <TextField
-                            label="Surface accès public"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.publicAccessSurface || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'publicAccessSurface',
-                              parseFloat(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0, step: 0.1 }}
-                            fullWidth
-                            slotProps={{
-                              input: {
-                                endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                              }
-                            }}
-                          />
-                        </Box>
-                      </Paper>
-                    ))}
-                  </Box>
-                </Box>
-
-                {/* Mobile View */}
-                <Box sx={{ display: { xs: 'block', md: 'none' } }}>
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                    {levelAssignments.map((assignment) => (
-                      <Paper key={assignment.levelNumber} variant="outlined" sx={{ p: 2 }}>
-                        <Typography variant="subtitle1" fontWeight="medium" sx={{ mb: 2 }}>
-                          Niveau {assignment.levelNumber}
-                        </Typography>
-
-                        <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                          <FormControl size="small" fullWidth>
-                            <InputLabel>Étage de bâtiment</InputLabel>
-                            <Select
-                              label="Étage de bâtiment"
-                              value={assignment.buildingFloorId || ''}
-                              onChange={(e) => updateLevelAssignment(
-                                assignment.levelNumber,
-                                'buildingFloorId',
-                                e.target.value === '' ? null : Number(e.target.value)
-                              )}
-                              displayEmpty
-                            >
-                              <MenuItem value="">Non affecté</MenuItem>
-                              {getAvailableBuildingFloors(formData.buildingId).map((floor) => (
-                                <MenuItem key={floor.id} value={floor.id}>
-                                  {floor.name}
-                                </MenuItem>
-                              ))}
-                            </Select>
-                          </FormControl>
-
-                          <TextField
-                            label="Nom de l'étage de partie"
-                            size="small"
-                            value={assignment.partFloorData?.name || ''}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'name',
-                              e.target.value
-                            )}
-                            placeholder={`Niveau ${assignment.levelNumber}`}
-                            fullWidth
-                          />
-
-                          <Box sx={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 2 }}>
-                            <TextField
-                              label="Niveau de partie"
-                              size="small"
-                              type="number"
-                              value={assignment.partFloorData?.levelNumber || assignment.levelNumber}
-                              onChange={(e) => updatePartFloorData(
-                                assignment.levelNumber,
-                                'levelNumber',
-                                parseInt(e.target.value) || assignment.levelNumber
-                              )}
-                              inputProps={{ min: 0 }}
-                              fullWidth
-                            />
-
-                            <TextField
-                              label="Public"
-                              size="small"
-                              type="number"
-                              value={assignment.partFloorData?.publicCount || 0}
-                              onChange={(e) => updatePartFloorData(
-                                assignment.levelNumber,
-                                'publicCount',
-                                parseInt(e.target.value) || 0
-                              )}
-                              inputProps={{ min: 0 }}
-                              fullWidth
-                            />
-                          </Box>
-
-                          <TextField
-                            label="Personnel"
-                            size="small"
-                            type="number"
-                            value={assignment.partFloorData?.staffCount || 0}
-                            onChange={(e) => updatePartFloorData(
-                              assignment.levelNumber,
-                              'staffCount',
-                              parseInt(e.target.value) || 0
-                            )}
-                            inputProps={{ min: 0 }}
-                            fullWidth
-                          />
-
-                          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                            <TextField
-                              label="Surface exploitation"
-                              size="small"
-                              type="number"
-                              value={assignment.partFloorData?.exploitationSurface || 0}
-                              onChange={(e) => updatePartFloorData(
-                                assignment.levelNumber,
-                                'exploitationSurface',
-                                parseFloat(e.target.value) || 0
-                              )}
-                              inputProps={{ min: 0, step: 0.1 }}
-                              fullWidth
-                              slotProps={{
-                                input: {
-                                  endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                                }
-                              }}
-                            />
-
-                            <TextField
-                              label="Surface GLA"
-                              size="small"
-                              type="number"
-                              value={assignment.partFloorData?.glaSurface || 0}
-                              onChange={(e) => updatePartFloorData(
-                                assignment.levelNumber,
-                                'glaSurface',
-                                parseFloat(e.target.value) || 0
-                              )}
-                              inputProps={{ min: 0, step: 0.1 }}
-                              fullWidth
-                              slotProps={{
-                                input: {
-                                  endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                                }
-                              }}
-                            />
-
-                            <TextField
-                              label="Surface accès public"
-                              size="small"
-                              type="number"
-                              value={assignment.partFloorData?.publicAccessSurface || 0}
-                              onChange={(e) => updatePartFloorData(
-                                assignment.levelNumber,
-                                'publicAccessSurface',
-                                parseFloat(e.target.value) || 0
-                              )}
-                              inputProps={{ min: 0, step: 0.1 }}
-                              fullWidth
-                              slotProps={{
-                                input: {
-                                  endAdornment: <InputAdornment position="end">m²</InputAdornment>
-                                }
-                              }}
-                            />
-                          </Box>
-                        </Box>
-                      </Paper>
-                    ))}
-                  </Box>
-                </Box>
-              </Box>
-            )}
-          </Box>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={closeDialog}>Annuler</Button>
-          <Button onClick={handleSubmit} variant="contained">
-            {dialogMode === 'create' ? 'Créer' : 'Modifier'}
-          </Button>
-        </DialogActions>
-      </Dialog>
+        onClose={closeDialog}
+        mode={dialogMode}
+        part={selectedPart}
+        buildings={buildings}
+        buildingFloors={buildingFloors}
+        includeDeleted={filters.includeDeleted}
+        onSubmit={handleDialogSubmit}
+      />
     </Box>
   );
 };
